+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added percentage calculation field to calculate the difference in percentage between two fields.
 * Fixed issue with saving default filter values.
 * Fixed issue with filter for multiple select fields.
+* Added filter for month.
 
 # Version 1.14.2